Move IProcessRunner and associated types to Buildvana.Core.Process
