+++ src/Buildvana.Tool/Services/Versioning/VersionService.cs
@@ -3,6 +3,7 @@
 
 using Buildvana.Core;
 using Buildvana.Core.Json;
+using Buildvana.Core.Process;
 using Buildvana.Tool.Services.Git;
 using Buildvana.Tool.Services.PublicApiFiles;
 using CommunityToolkit.Diagnostics;
